requirements.txt
Flask==1.1.2
Keras==2.3.1
tensorflow==1.14.0
matplotlib==3.1.3
numpy==1.16.4
scikit_image==0.16.2
h5py==2.8.0
scipy==1.4.1
imgaug
ipython==7.15.0
Pillow
skimag
